Incorporate existing UI components into Search UI (#357)
* Adjust AdminScreen and Modal for TwoColumnsLayout integration (#351)

* Updates on TwoColumnsLayout, fix regressions on AdminScreen and Modal and upgrade other layout stories

* Fix height issue on AdminScreen

* Remove out of the scope changes and include more stories for TwoColumnsLayout

* Include TwoColumnsLayout showcase composed with AdminScreens and Modal, revert ListsScreen and SearchForm stories updates

* Fix height concerns for AdminScreen and TwoColumnsLayout and improve showcases of ListsScreen

* Cleanup code to pass linter

* Update AdminScreen

* Fixes Modal at AdminScreen

* Update AdminScreen and Modal visibility handling

* Fix typo

* Rename Tab to GenericTabs and Collapsible to GenericCollapsible (#353)

* Rename Tab to GenericTabs

* Rename Collapsible to GenericCollapsible

* Update snapshots

* Handle state changes (#356)

* Generic tabs fix

* Add debug test story

* Make a consistent story name

# Conflicts:
#	lib/auto-ui.js.map
