+++ src/components/ui/AdminScreen.js
@@ -10,14 +10,24 @@ const cmz = require('cmz')
 
 const cx = {
   admin: cmz(`
-    height: 100%
+    height: 100vh
+    min-height: 100vh
+    overflow-y: hidden
     position: relative
+    display: flex
+    flex-direction: column
   `),
 
   header: cmz(`
+    width: 100%
     height: 56px
-    position: relative
     z-index: 10000
+  `),
+
+  content: cmz(`
+    flex: 1 0 0
+    overflow: auto
+    height: 100%
   `)
 }
 
@@ -25,8 +35,8 @@ type Props = {
   header?: Element<*>,
   children?: Element<*>,
   modal?: {
-    onClose: Function,
-    content: Element<*>
+    onClose?: Function,
+    content?: Element<*>
   }
 }
 
@@ -39,17 +49,22 @@ class AdminScreen extends PureComponent<Props, void> {
     }
   }
 
+  handleModalClose = () => {
+    const { modal } = this.props
+    modal && modal.onClose && modal.onClose()
+  }
+
   render () {
     const { header, children, modal } = this.props
     return (
       <div className={cx.admin}>
         {modal && modal.content && (
-          <Modal onClose={modal.onClose}>
+          <Modal onClose={this.handleModalClose}>
             {modal.content}
           </Modal>
         )}
         <div className={cx.header}>{header}</div>
-        {children}
+        <div className={cx.content}>{children}</div>
       </div>
     )
   }
