--- src/components/ui/Collapsible/index.js
@@ -1,9 +1,0 @@
-import Body from './Body'
-import Container from './Container'
-import Header from './Header'
-
-export default {
-  Body,
-  Container,
-  Header
-}
