+++ src/components/ui/GenericCollapsible/index.js
@@ -1,0 +1,9 @@
+import Body from './Body'
+import Container from './Container'
+import Header from './Header'
+
+export default {
+  Body,
+  Container,
+  Header
+}
