+++ src/components/ui/GenericTabs/index.js
@@ -1,0 +1,11 @@
+// @flow
+
+import Pane from './Pane'
+import Head from './Head'
+import Container from './Container'
+
+export default {
+  Container,
+  Head,
+  Pane
+}
