+++ src/components/ui/Modal.js
@@ -1,8 +1,10 @@
 // @flow
 
-import React, { Component } from 'react'
+import React, { PureComponent } from 'react'
 
 import SvgIcon from './SvgIcon'
+
+import theme from '../../styles/theme'
 
 import type { Element } from 'react'
 
@@ -13,68 +15,50 @@ type Props = {
   children?: Element<*>|string
 }
 
-type State = {
-  open: boolean
-}
-
 const cx = {
   modal: cmz(`
-    & {
-      background: rgba(0, 0, 0, .3)
-      height: 100%
-      width: 100%
-      position: absolute
-      display: flex
-      overflow: auto
-      padding: 30px
-      box-sizing: border-box
-      z-index: 99999
-      opacity: 0
-      visibility: hidden
-      outline: none
-    }
-
-    &.open {
-      opacity: 1
-      visibility: visible
-    }
+    background: rgba(0, 0, 0, .3)
+    height: 100%
+    width: 100%
+    position: absolute
+    display: flex
+    overflow: auto
+    padding: 30px
+    box-sizing: border-box
+    z-index: 99999
+    outline: none
   `),
 
   frame: cmz(`
     position: relative
-    background: white
+    background: ${theme.baseBright}
     margin: auto
     max-width: 100%
     min-width: 38px
     min-height: 40px
   `),
 
-  close: cmz('closemodal', `
+  content: cmz(`
+    overflow-x: auto
+  `),
+
+  close: cmz(`
     position: absolute
     right: 12px
     top: 12px
     cursor: pointer
+    z-index: 1
   `)
 }
 
-class Modal extends Component<Props, State> {
-  state = {
-    open: false
-  }
-
-  componentDidMount () {
-    this.setState({ open: true })
-  }
-
+class Modal extends PureComponent<Props, void> {
   noClick = (event: any) => {
     event && event.stopPropagation()
   }
 
   handleClose = () => {
     const { onClose } = this.props
-    this.setState({ open: false }, () => {
-      onClose && onClose()
-    })
+    onClose && onClose()
   }
 
   handleKeyPress = (e: any) => {
@@ -89,20 +73,20 @@ class Modal extends Component<Props, State> {
 
   render () {
     const { children } = this.props
-    const modalClassName = [cx.modal, this.state.open && 'open'].join(' ')
-
     return children ? (
       <div
-        className={modalClassName}
+        className={cx.modal}
         onClick={this.handleClose}
         onKeyDown={this.handleKeyPress}
         tabIndex={0}
       >
         <section className={cx.frame} onClick={this.noClick}>
-          {children}
           <a className={cx.close} onClick={this.handleClose}>
             <SvgIcon icon='x' color='grayscale' />
           </a>
+          <div className={cx.content}>
+            {children}
+          </div>
         </section>
       </div>
     ) : null
