+++ src/components/ui/Modal.stories.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
+import State from '../../utils/State'
 
 import Modal from './Modal'
 import ListsEditor from './ListsEditor'
@@ -11,55 +12,78 @@ const itemsArray = Array(18).fill({}).map((item, i) => ({
   archived: i > 12 && true
 }))
 
+const sampleModalContent = [
+  <h1 key='h1'>Modal content goes here</h1>,
+  ...Array(5).fill('Anything goes in the modal content body').map((each, i) => <div key={`content-${i}`}>{each}</div>)
+]
+
+const sampleModalContentTall = [
+  <h1 key='h1'>Modal content goes here</h1>,
+  ...Array(80).fill('Anything goes in the modal content body').map((each, i) => <div key={`content-${i}`}>{each}</div>)
+]
+
+const sampleModalContentLong = [
+  <h1 key='h1'>Modal content goes here</h1>,
+  ...Array(5).fill('AnythinggoesinthemodalcontentbodyAnythinggoesinthemodalcontentbodyAnythinggoesinthemodalcontentbodyAnythinggoesinthemodalcontentbodyAnythinggoesinthemodalcontentbodyAnythinggoesinthemodalcontentbodyAnythinggoesinthemodalcontentbodyAnythinggoesinthemodalcontentbodyAnythinggoesinthemodalcontentbody').map((each, i) => <div key={`content-${i}`}>{each}</div>)
+]
+
+const Body = ({ children }) => (
+  <div style={{ minHeight: '100vh', overflow: 'hidden' }}>
+    <style dangerouslySetInnerHTML={{ __html: `
+      body { margin: 0; }
+    ` }} />
+    {children}
+  </div>
+)
+
+const StoryModal = props => (
+  <State initialState={{ isOpen: true }}>
+    {({ setState, state }) => state.isOpen ? (
+      <Modal
+        onClose={() => setState({ isOpen: false })}
+        {...props}
+      />
+    ) : null}
+  </State>
+)
+
 storiesOf('UI Components/Modal', module)
   .add('basic modal', () => (
-    <div style={{ position: 'relative', minHeight: '500px' }}>
-      <Modal>
-        <h1>Modal content goes here</h1>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-      </Modal>
-    </div>
+    <Body>
+      <StoryModal>
+        {sampleModalContent}
+      </StoryModal>
+    </Body>
   ))
-  .add('modal with content bigger than viewport', () => (
-    <div style={{ position: 'relative', minHeight: '300px' }}>
-      <Modal>
-        <h1>Modal content goes here</h1>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-        <p>lorem ipsum lorem ipsum lorem ipsum lorem ipsum lorem ipsum</p>
-      </Modal>
-    </div>
+  .add('modal with content taller than viewport', () => (
+    <Body>
+      <StoryModal>
+        {sampleModalContentTall}
+      </StoryModal>
+    </Body>
   ))
-  .add('ListEditor use case', () => (
-    <div style={{ position: 'relative', minHeight: '800px' }}>
-      <Modal>
+  .add('modal with content longer than viewport', () => (
+    <Body>
+      <StoryModal>
+        {sampleModalContentLong}
+      </StoryModal>
+    </Body>
+  ))
+  .add('modal with no content', () => (
+    <Body>
+      <StoryModal />
+    </Body>
+  ))
+  .add('missing props (does component explode?)', () => (
+    <Body>
+      <Modal />
+    </Body>
+  ))
+
+storiesOf('UI Components/Modal/Use cases', module)
+  .add('with ListEditor', () => (
+    <Body>
+      <StoryModal>
         <ListsEditor
           collectionLabel='List'
           title='Lists'
@@ -69,11 +93,6 @@ storiesOf('UI Components/Modal', module)
           onDelete={action('onDelete')}
           onCreateNew={action('onCreateNew')}
         />
-      </Modal>
-    </div>
-  ))
-  .add('missing props (does component explode?)', () => (
-    <div style={{ position: 'relative', minHeight: '100px' }}>
-      <Modal />
-    </div>
+      </StoryModal>
+    </Body>
   ))
