--- src/components/ui/Tab/index.js
@@ -1,11 +1,0 @@
-// @flow
-
-import Pane from './Pane'
-import Head from './Head'
-import Container from './Container'
-
-export default {
-  Container,
-  Head,
-  Pane
-}
